Organize actions by subject
